skill

+++ backend/tools/SkillCraft.Seeding/Krakenar/Tasks/SeedContentTypesTask.cs
@@ -8,6 +8,12 @@ namespace SkillCraft.Seeding.Krakenar.Tasks;
 internal class SeedContentTypesTask : SeedingTask
 {
   public override string? Description => "Seeds the Content Types into Krakenar.";
+  public bool FieldDefinitions { get; }
+
+  public SeedContentTypesTask(bool fieldDefinitions = false)
+  {
+    FieldDefinitions = fieldDefinitions;
+  }
 }
 
 internal class SeedContentTypesTaskHandler : INotificationHandler<SeedContentTypesTask>
@@ -26,7 +32,7 @@ internal class SeedContentTypesTaskHandler : INotificationHandler<SeedContentTyp
     _logger = logger;
   }
 
-  public async Task Handle(SeedContentTypesTask _, CancellationToken cancellationToken)
+  public async Task Handle(SeedContentTypesTask task, CancellationToken cancellationToken)
   {
     string json = await File.ReadAllTextAsync("Krakenar/data/content_types.json", Encoding.UTF8, cancellationToken);
     IEnumerable<ContentTypePayload>? payloads = SeedingSerializer.Deserialize<IEnumerable<ContentTypePayload>>(json);
@@ -34,24 +40,31 @@ internal class SeedContentTypesTaskHandler : INotificationHandler<SeedContentTyp
     {
       foreach (ContentTypePayload payload in payloads)
       {
-        CreateOrReplaceContentTypeResult result = await _contentTypeService.CreateOrReplaceAsync(payload, payload.Id, version: null, cancellationToken);
-        ContentType contentType = result.ContentType ?? throw new InvalidOperationException($"'ContentTypeService.CreateOrReplaceAsync' returned null for content type 'Id={payload.Id}'.");
-        _logger.LogInformation("The content type '{ContentType}' was {Action}.", contentType.DisplayName ?? contentType.UniqueName, result.Created ? "created" : "replaced");
-
-        HashSet<Guid> fieldIds = payload.Fields.Select(x => x.Id).ToHashSet();
-        foreach (FieldDefinition field in contentType.Fields)
+        if (task.FieldDefinitions)
         {
-          if (!fieldIds.Contains(field.Id))
+          ContentType contentType = await _contentTypeService.ReadAsync(payload.Id, uniqueName: null, cancellationToken)
+            ?? throw new InvalidOperationException($"The content type 'Id={payload.Id}' was not found.");
+          HashSet<Guid> fieldIds = payload.Fields.Select(x => x.Id).ToHashSet();
+          foreach (FieldDefinition field in contentType.Fields)
           {
-            await _fieldDefinitionService.DeleteAsync(contentType.Id, field.Id, cancellationToken);
-            _logger.LogInformation("The field definition '{FieldDefinition}' was deleted.", field.DisplayName ?? field.UniqueName);
+            if (!fieldIds.Contains(field.Id))
+            {
+              await _fieldDefinitionService.DeleteAsync(contentType.Id, field.Id, cancellationToken);
+              _logger.LogInformation("The field definition '{FieldDefinition}' was deleted.", field.DisplayName ?? field.UniqueName);
+            }
+          }
+
+          foreach (FieldDefinitionPayload field in payload.Fields)
+          {
+            await _fieldDefinitionService.CreateOrReplaceAsync(contentType.Id, field, field.Id, cancellationToken);
+            _logger.LogInformation("The field definition '{FieldDefinition}' was saved.", field.DisplayName ?? field.UniqueName);
           }
         }
-
-        foreach (FieldDefinitionPayload field in payload.Fields)
+        else
         {
-          await _fieldDefinitionService.CreateOrReplaceAsync(contentType.Id, field, field.Id, cancellationToken);
-          _logger.LogInformation("The field definition '{FieldDefinition}' was saved.", field.DisplayName ?? field.UniqueName);
+          CreateOrReplaceContentTypeResult result = await _contentTypeService.CreateOrReplaceAsync(payload, payload.Id, version: null, cancellationToken);
+          ContentType contentType = result.ContentType ?? throw new InvalidOperationException($"'ContentTypeService.CreateOrReplaceAsync' returned null for content type 'Id={payload.Id}'.");
+          _logger.LogInformation("The content type '{ContentType}' was {Action}.", contentType.DisplayName ?? contentType.UniqueName, result.Created ? "created" : "replaced");
         }
       }
     }
